requirements.txt
Pillow==10.2.0
PyYAML==6.0.1
numpy==1.24.3
psutil==5.9.5
requests==2.31.0